refactor: migrate GitHub app code to Firebase functions
- Removed the existing GitHub app implementation from openci-runner/github-apps.
- Added new implementation in openci-runner/firebase/functions/probot.
- Updated test fixtures and test cases to reflect the new structure.
- Deleted obsolete files related to the previous implementation.
