--- .github/workflows/reusable-firebase-functions-ci.yml
@@ -1,31 +1,0 @@
-name: Firebase Functions Tests
-
-on:
-  workflow_call:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: openci-runner/firebase/functions
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version-file: .nvmrc
-
-      - name: Install dependencies for GitHub apps
-        working-directory: openci-runner/github-apps
-        run: |
-          npm ci
-          npm run build
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Check build
-        run: npm run build
