--- .github/workflows/reusable-github-apps-ci.yml
@@ -1,28 +1,0 @@
-name: GitHub Apps Tests
-
-on:
-  workflow_call:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    defaults:
-      run:
-        working-directory: openci-runner/github-apps
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Set up Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version-file: .nvmrc
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Run tests
-        run: npm test
-
-      - name: Check build
-        run: npm run build
