+++ openci-runner/firebase/functions/test/fixtures/issues.opened.json
