--- openci-runner/github-apps/.dockerignore
@@ -1,12 +1,0 @@
-**/node_modules/
-**/.git
-**/README.md
-**/LICENSE
-**/.vscode
-**/npm-debug.log
-**/coverage
-**/.env
-**/.editorconfig
-**/dist
-**/*.pem
-Dockerfile
