--- openci-runner/github-apps/.env.example
@@ -1,9 +1,0 @@
-# The ID of your GitHub App
-APP_ID=
-WEBHOOK_SECRET=development
-
-# Use `trace` to get verbose logging or `info` to show less
-LOG_LEVEL=debug
-
-# Go to https://smee.io/new set this to the URL that you are redirected to.
-WEBHOOK_PROXY_URL=
