--- openci-runner/github-apps/.gitignore
@@ -1,7 +1,0 @@
-node_modules
-npm-debug.log
-*.pem
-!mock-cert.pem
-.env
-coverage
-lib
