--- openci-runner/github-apps/Dockerfile
@@ -1,8 +1,0 @@
-FROM node:20-slim
-WORKDIR /usr/src/app
-COPY package.json package-lock.json ./
-RUN npm ci --production
-RUN npm cache clean --force
-ENV NODE_ENV="production"
-COPY . .
-CMD [ "npm", "start" ]
