--- openci-runner/github-apps/README.md
@@ -1,33 +1,0 @@
-# github-apps
-
-> A GitHub App built with [Probot](https://github.com/probot/probot) that A OpenCI Runner App
-
-## Setup
-
-```sh
-# Install dependencies
-npm install
-
-# Run the bot
-npm start
-```
-
-## Docker
-
-```sh
-# 1. Build container
-docker build -t github-apps .
-
-# 2. Start container
-docker run -e APP_ID=<app-id> -e PRIVATE_KEY=<pem-value> github-apps
-```
-
-## Contributing
-
-If you have suggestions for how github-apps could be improved, or want to report a bug, open an issue! We'd love all and any contributions.
-
-For more, check out the [Contributing Guide](CONTRIBUTING.md).
-
-## License
-
-[ISC](LICENSE) © 2025 mafreud
