--- openci-runner/github-apps/package.json
@@ -1,33 +1,0 @@
-{
-	"name": "github-apps",
-	"version": "1.0.0",
-	"private": true,
-	"description": "An OpenCI Runner App",
-	"author": "OpenCI inc",
-	"license": "ISC",
-	"homepage": "https://openci.run",
-	"keywords": ["probot", "github", "probot-app"],
-	"scripts": {
-		"build": "tsc",
-		"start": "probot run ./lib/index.js",
-		"test": "vitest"
-	},
-	"dependencies": {
-		"probot": "^14.0.2"
-	},
-	"devDependencies": {
-		"@types/node": "^24.6.2",
-		"@vitest/coverage-v8": "^3.2.4",
-		"@vitest/ui": "^3.2.4",
-		"@vitest/utils": "^3.2.4",
-		"nock": "^14.0.10",
-		"smee-client": "^4.3.1",
-		"typescript": "^5.9.3",
-		"vite": "^7.1.9",
-		"vitest": "^3.2.4"
-	},
-	"engines": {
-		"node": "^22.0.0"
-	},
-	"type": "module"
-}
