--- openci-runner/github-apps/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "incremental": true,
-    "target": "es2022",
-    "module": "es2022",
-    "declaration": true,
-    "sourceMap": true,
-    "outDir": "./lib",
-    /* Strict Type-Checking Options */
-    "strict": true,
-    /* Additional Checks */
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "moduleResolution": "bundler",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    // TODO(someone): remove this
-    "skipLibCheck": true,
-  },
-  "include": [
-    "src/"
-  ],
-  "compileOnSave": true
-}
