--- openci-runner/github-apps/vitest.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-	test: {
-		include: ["test/**/*.test.ts"],
-		coverage: {
-			provider: "v8",
-		},
-	},
-});
